requirements.txt
Flask==2.3.2
APScheduler==3.10.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
